why?!!!1111!!!11!1

+++ 1.py
@@ -2,5 +2,5 @@ a = int(input())
 b = int(input())
 
 c = a^(2*b) - b^(2*a)
-c = float(c)
+c = int(c)
 print(c)
